+++ rest-service/MusicLibrary/src/main/java/mfsi/learnmvc/repository/PlaylistRepository.java
@@ -6,9 +6,8 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 import mfsi.learnmvc.domain.Playlist;
-import mfsi.learnmvc.domain.Singer;
 
 @Repository
-public interface PlaylistRepository extends CrudRepository<Playlist, Integer> {
+public interface PlaylistRepository extends CrudRepository<Playlist, Long> {
 	public List<Playlist> findAll();
 }
