+++ rest-service/MusicLibrary/src/main/java/mfsi/learnmvc/rest/SingerController.java
@@ -1,7 +1,5 @@
 package mfsi.learnmvc.rest;
 
-import java.util.List;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -12,17 +10,16 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-import mfsi.learnmvc.domain.Singer;
 import mfsi.learnmvc.dto.SingerDto;
 import mfsi.learnmvc.service.SingerService;
 
 @RestController
 @RequestMapping("/singer")
 public class SingerController {
-	
+
 	@Autowired
 	private SingerService service;
-	
+
 	@RequestMapping(value = "/all", method = RequestMethod.GET)
 	public ResponseEntity<?> getAll() {
 		ResponseEntity<?> response = new ResponseEntity<>(service.getAll(), HttpStatus.OK);
@@ -34,17 +31,16 @@ public class SingerController {
 		ResponseEntity<?> response = new ResponseEntity<>(service.save(dto), HttpStatus.OK);
 		return response;
 	}
-	
+
 	@RequestMapping(value = "/delete/{id}", method = RequestMethod.DELETE)
-	public void delete(@PathVariable Integer id) {
+	public void delete(@PathVariable Long id) {
 		service.delete(id);
 	}
-	
+
 	@RequestMapping(value = "/search", method = RequestMethod.GET)
-	public ResponseEntity<?> search(@RequestParam String keyword ){
+	public ResponseEntity<?> search(@RequestParam String keyword) {
 		ResponseEntity<?> response = new ResponseEntity<>(service.search(keyword), HttpStatus.OK);
 		return response;
 	}
-
 
 }
